Move testBindExceptionMessageFormatting to integration tests.
